# swagger/ft.yaml
openapi: 3.0.0
info:
  title: ftx
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
...
  /api/video/:
    put:
      tags: [ video ]
      summary: Upload a video
      operationId: put-api-video
      description: Upload local video to videoservice
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/Video'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unvideoized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
  /api/video/{videoId}:
    post:
      tags: [ video ]
      summary: Update video information
      operationId: post-api-video-{videoId}
      description: Update video information
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "videoId"
          in: 'path'
          description: "videoId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/Video'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unvideoized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
    get:
      tags: [ video ]
      summary: Fetch the video
      operationId: get-api-video-{videoId}
      description: Fetch the video by Id
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "videoId"
          in: 'path'
          description: "videoId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/Video'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unvideoized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
  /api/video/{videoId}/like:
    get:
      tags: [ video ]
      summary: Like the video
      operationId: get-api-video-{videoId}-like
      description: Like the video by Id
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "videoId"
          in: 'path'
          description: "videoId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/Video'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unvideoized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
  /api/video/{videoId}/dislike:
    get:
      tags: [ video ]
      summary: Dislike the video
      operationId: get-api-video-{videoId}-dislike
      description: Diskike the video by Id
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "videoId"
          in: 'path'
          description: "videoId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/Video'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unvideoized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
  /api/video/{videoId}/viewmark:
    post:
      tags: [ video ]
      summary: Update video viewmark information
      operationId: post-api-video-{videoId}-viewmark
      description: Update video viewmark information by Id
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "videoId"
          in: 'path'
          description: "videoId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/Video'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unvideoized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
    get:
      tags: [ video ]
      summary: Fetch video viewmark
      operationId: get-api-video-{videoId}-viewmark
      description: Fetch video viewmark by Id
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "videoId"
          in: 'path'
          description: "videoId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/Video'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unvideoized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
  /api/video/search:
    get:
      tags: [ video ]
      summary: Searching for videos
      operationId: get-api-video-search
      description: Searching for videos
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "videoization token"
          required: true
          schema:
            type: "string"
        - name: "q"
          in: 'query'
          description: "Query string for searching"
          schema:
            type: "string"
        - name: "author"
          in: 'query'
          description: "Author Id for searching"
          schema:
            type: "string"
        - name: "tags"
          in: 'query'
          description: "Tag describes group of topic related video"
          schema:
            type: array
            items:
              type: string
        - name: 'pageSize'
          in: 'query'
          description: "Maximum number of items per page after the search"
          schema:
            type: "string"
      responses:
        '200':
          description: Searching for videos
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Video'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unvideoized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
...
  '/api/pitch/{pitchId}/video':
    post:
      tags: [ pitch ]
      summary: upload video for pitch
      operationId: post-api-pitch-{pitchId}-video
      description: upload video for pitch
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: pitchId
          in: 'path'
          description: "pitchId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Pitch'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: []
...
components:
...
  schemas:
...
    Video:
      properties:
        title:
          type: string
        description:
          type: string
        previewImageUrl:
          type: string
        tags:
          type: string
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
        likes:
          type: array
          items:
            $ref: '#/components/schemas/UserProfileUserEditable'
        likesCount:
          type: number
        accessGroup:
          type: array
          items:
            type: string
            enum:
              - 'PREMIUM'
    Author:
      properties:
        name:
          type: string
        surname:
          type: string
        title:
          type: string
        description:
          type: string
        avatarUrl:
          type: string
...
    Pitch:
      properties:
        businessIedaId:
          type: object
          $ref: '#/components/schemas/BusinessIdeaId'
        userId:
          type: object
          $ref: '#/components/schemas/UserProfileId'
        title:
          type: string
          required:
            - "can't be blank"
        description:
          type: string
        language:
          type: string
        version:
          type: number
        tags:
          type: array
          items:
            type: string
        active:
          type: boolean
        deleted:
          type: boolean
        private:
          type: boolean
        avgRate:
          type: number
          default: 0
        reviewFeedback:
          type: object
          $ref: '#/components/schemas/ReviewFeedback'
        reviewed:
          type: boolean
          default: false
        rejected:
          type: boolean
          default: false
        reviewedBy:
          type: object
          $ref: '#/components/schemas/UserProfileId'
        reviews:
          type: array
          $ref: '#/components/schemas/PitchReview'
        reviewCount:
          type: number
          default: 0
        localVideoPath:
          type: string
        brightcove:
          type: object
        video:
          type: object
          $ref: '#/components/schemas/PitchVideo'
        likes:
          type: number
          description: count of Like when user touch like button on pitch.It does not send this field if there are not data.
          example: 2
        views:
          type: number
          description: count of view when user access on this pitch..It does not send this field if data is nothing.
          example: 5
    PitchReview:
      properties:
        feedback:
          type: string
          required:
            - "can't be blank"
        avgRate:
          type: number
          default: 0
        rate:
          type: array
          items:
            $ref: '#/components/schemas/PitchRate'
        userId:
          type: object
          $ref: '#/components/schemas/UserProfileId'
        submitDate:
          type: string
          format: date
    PitchRate:
      properties:
        reviewCategoryId:
          type: object
          $ref: '#/components/schemas/ReviewCategoryId'
        reviewRating:
          type: number
          minimum: 0
          maximum: 10
    ReviewFeedback:
      properties:
        userId:
          type: object
          $ref: '#/components/schemas/UserProfileId'
        rate:
          type: number
          enum:
            - -1
            - 1
        date:
          type: string
          format: date
    PitchVideo:
      properties:
        thumbnailUrl:
          type: string
        posterUrl:
          type: string
        videoUrl:
          type: string
        updateDate:
          type: string
          format: date
...
    BusinessIdeaId:
      description: ''
      type: object
      x-examples:
        BusinessIdeaId example:
          id: string
      title: BusinessIdeaId
      properties:
        id:
          type: string
          minLength: 1
          description: |
            Unique business-idea ID
          example: 6fc52fbcf219cd786fabffab
      required:
        - id
...
    ReviewCategoryId:
      description: ''
      type: object
      x-examples:
        ReviewCategoryId example:
          id: string
      title: ReviewCategoryId
      properties:
        id:
          type: string
          minLength: 1
          description: |
            Unique review-category ID
          example: cac52fbcf307ad786fabf261
      required:
        - id
    UserProfileId:
      description: ''
      type: object
      x-examples:
        UserProfileId example:
          id: string
      title: UserProfileId
      properties:
        id:
          type: string
          minLength: 1
          description: |
            Unique user ID
          example: 5fc52fbcf779cd786fabfff0
      required:
        - id
...
    UserProfileUserEditable:
      description: ''
      type: object
      x-examples:
        example-1:
          email: string
          name: string
          surname: string
          avatarUrl: string
          bio: string
          theme: string
          watchedVideos: 0
      title: UserProfileUserEditable
      properties:
        email:
          type: string
          minLength: 1
          description: User email. Should be unique across the users
          example: test@domain.com
        name:
          type: string
          minLength: 1
          description: User first name
          example: BoJack
        surname:
          type: string
          minLength: 1
          description: User last name
          example: Horseman
        avatarUrl:
          type: string
          minLength: 1
          description: User's avatar URL
          example: 'https://i.pravatar.cc/300'
        bio:
          type: string
          minLength: 1
          description: User's bio
          example: User's bio goes here
        theme:
          type: string
          minLength: 1
          description: User's client application theme
          enum:
            - light
            - dark
...
    GenericErrors:
      title: GenericErrors
      type: object
      x-examples:
        example-1:
          - string
      properties:
        errors:
          type: array
          description: Errors list occurred during request processing
          items:
            type: string
      description: ''
...
  securitySchemes:
    API Key:
      name: API Key
      type: apiKey
      in: header
      description: Firebase token